requirements.txt
Django==5.0.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
requests==2.31.0
google-generativeai==0.3.0
beautifulsoup4==4.12.2
pytube==15.0.0
youtube-transcript-api==0.6.3
gunicorn==21.2.0
python-decouple==3.8
